+++ android/app/src/main/java/com/storybooknative/stubs/StorybookStubPackage.kt
@@ -13,6 +13,7 @@ class StorybookStubPackage : ReactPackage {
             AuthorProfileEvents(),
             ReactAnalytics(),
             ReactConfig(),
+            SectionEvents(),
             TopicEvents()
     )