+++ packages/article-magazine-standard/src/article-magazine-standard.js
@@ -26,7 +26,7 @@ class ArticleMagazineStandard extends Component {
   }
 
   renderHeader({ width }) {
-    const { article, onAuthorPress, onVideoPress } = this.props;
+    const { article, onAuthorPress, onImagePress, onVideoPress } = this.props;
     const {
       byline,
       expirableFlags,
@@ -58,6 +58,7 @@ class ArticleMagazineStandard extends Component {
             <LeadAsset
               {...getLeadAsset(article)}
               getImageCrop={getStandardTemplateCrop}
+              onImagePress={onImagePress}
               onVideoPress={onVideoPress}
               renderCaption={({ caption }) => <CentredCaption {...caption} />}
               style={[
@@ -92,6 +93,7 @@ class ArticleMagazineStandard extends Component {
       onAuthorPress,
       onCommentGuidelinesPress,
       onCommentsPress,
+      onImagePress,
       onLinkPress,
       onRelatedArticlePress,
       onTopicPress,
@@ -111,6 +113,7 @@ class ArticleMagazineStandard extends Component {
         onAuthorPress={onAuthorPress}
         onCommentGuidelinesPress={onCommentGuidelinesPress}
         onCommentsPress={onCommentsPress}
+        onImagePress={onImagePress}
         onLinkPress={onLinkPress}
         onRelatedArticlePress={onRelatedArticlePress}
         onTopicPress={onTopicPress}