+++ packages/article-main-comment/src/article-prop-types/article-prop-types.base.js
@@ -12,13 +12,15 @@ const articlePagePropTypes = {
     })
   }),
   isLoading: PropTypes.bool,
+  onImagePress: PropTypes.func,
   receiveChildList: PropTypes.func.isRequired
 };
 
 const articlePageDefaultProps = {
   article: null,
   error: null,
-  isLoading: false
+  isLoading: false,
+  onImagePress: null
 };
 
 export { articlePagePropTypes, articlePageDefaultProps };