+++ packages/pages/pages.showcase.native.js
@@ -1,11 +1,8 @@
 /* eslint-disable react/prop-types */
 import React from "react";
-import { ActivityIndicator, Text } from "react-native";
 import { sections } from "@times-components/storybook";
-import { EditionProvider } from "@times-components/provider";
 import { Article, AuthorProfile, Topic } from "./src/pages";
-import Section from "./src/section/section";
-import withNativeProvider from "./src/with-native-provider";
+import Section from "./section-showcase-helper";
 
 export default {
   children: [
@@ -36,26 +33,9 @@ export default {
           "Edition id",
           "2b6e462c-225f-11e9-b782-40e94f317da5"
         );
-        const sectionTitle = sections[select("Section", sections, "News")];
+        const sectionTitle = select("Section", sections, "News");
 
-        return withNativeProvider(
-          <EditionProvider debounceTimeMs={0} id={editionId}>
-            {({ edition, error, isLoading }) => {
-              if (isLoading) {
-                return <ActivityIndicator size="large" />;
-              }
-              if (error) {
-                return <Text>{JSON.stringify(error)}</Text>;
-              }
-              const { publicationName: pubName } = edition;
-              return edition.sections
-                .filter(({ title }) => title === sectionTitle)
-                .map(sectionData => (
-                  <Section publicationName={pubName} section={sectionData} />
-                ));
-            }}
-          </EditionProvider>
-        );
+        return <Section editionId={editionId} sectionTitle={sectionTitle} />;
       },
       name: "Section",
       type: "story"