+++ packages/pagination/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [3.2.33](https://github.com/newsuk/times-components/compare/@times-components/pagination@3.2.32...@times-components/pagination@3.2.33) (2019-04-02)
+
+**Note:** Version bump only for package @times-components/pagination
+
+
+
+
+
 ## [3.2.32](https://github.com/newsuk/times-components/compare/@times-components/pagination@3.2.31...@times-components/pagination@3.2.32) (2019-03-14)
 
 **Note:** Version bump only for package @times-components/pagination