+++ packages/provider-test-tools/CHANGELOG.md
@@ -3,6 +3,17 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [1.15.0](https://github.com/newsuk/times-components/compare/@times-components/provider-test-tools@1.14.7...@times-components/provider-test-tools@1.15.0) (2019-04-03)
+
+
+### Features
+
+* gather data for image gallery ([#1835](https://github.com/newsuk/times-components/issues/1835)) ([52508b8](https://github.com/newsuk/times-components/commit/52508b8))
+
+
+
+
+
 ## [1.14.7](https://github.com/newsuk/times-components/compare/@times-components/provider-test-tools@1.14.6...@times-components/provider-test-tools@1.14.7) (2019-04-02)
 
 **Note:** Version bump only for package @times-components/provider-test-tools